Swapped A/B & X/Y in the Switch Backend
This change comes from how the firmware interacts with consoles like the PS4/5. Using a Brook Wingman, you can connect via holding X for Switch backend, but for some reason the other backends don't work.

Since A/B & X/Y are swapped on Switch controllers, I had to switch them in the backend to keep FGC Mode the same on PC & Console. With the new mode UltimateUSB, A&B are swapped so that when you plug in to Switch there's no noticeable change. Ultimate Mode is unchanged.